expose ModelDims unconditionally; forward spark-storage features
Two coupled changes that move spark-model closer to compiling on
Apple Silicon:

1. `ModelDims` was a plain POD struct (six u16/u32 fields, no GPU
   state) gated under `cfg(feature = "cuda")` because it lived in
   `high_speed_swap.rs`. Every layer's forward signature in
   spark-model threads `ModelDims` through, so this was the single
   biggest CUDA-only symbol leaking into otherwise-portable code.
   Moved to a new `spark_storage::model_dims` module that is
   always compiled. The cuda-only `high_speed_swap` module now
   `pub use crate::model_dims::ModelDims;` so existing callers
   keep working unchanged.

2. spark-model/Cargo.toml: thread spark-storage's `cuda` / `metal`
   features through spark-model's own feature flags, and add
   `default-features = false` to the path-dep line. Without this,
   spark-model would silently re-enable spark-storage's `cuda`
   default even when built under `--features metal` — exactly the
   pattern PR1 used for atlas-core, applied one layer further
   down the dep chain.

Verification: `cargo check -p spark-storage --no-default-features
--features metal` still passes; spark-model still has work to do
(~24 remaining errors at `with_local` / `local_installed`
consumer sites + one `atlas_core::device::sm121` use) which a
follow-on commit will address.
